+++ lib/phone_model.dart
@@ -1,0 +1,15 @@
+class PhoneModel {
+  final String name;
+  final String flag;
+  final String countryCode;
+  final String dialCode;
+  final String phoneNumber;
+
+  PhoneModel({
+    required this.name,
+    required this.flag,
+    required this.countryCode,
+    required this.dialCode,
+    required this.phoneNumber,
+  });
+}