+++ CHANGELOG.md
@@ -5,3 +5,8 @@
 ## 0.0.2
 
 * Flag size
+
+## 0.0.3
+* initial country code
+* picked country code 
+* provide all country information on picked